`@inheritDoc` now uses declaration references
